refactor(session-02): Updated attribute types after problems with endpoints.

+++ warehouse-management-api/warehouse-management-api/Contracts/UpdateProductPriceRequest.cs
@@ -1,3 +1,3 @@
 ﻿namespace warehouse_management_api.Contracts;
 
-public record UpdateProductPriceRequest(int Price);
+public record UpdateProductPriceRequest(decimal Price);
